backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.11
python-decouple==3.8
django-cors-headers==4.3.1
Pillow==10.4.0
django-cloudinary-storage==0.3.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.6
stripe==8.2.0
reportlab==4.0.9
gunicorn==21.2.0
whitenoise==6.6.0